Add files via upload
Below are java back end files to handle booking operation. BookARide can act as a Simple facade by providing a unifying interface to a set of interfaces in a subsystem. 

SQS can act as a messaging service to handle multiple requests of the client and it is implemented as a command pattern. We can add as multiple message queues based on the requests received from the clients.